remove old model, move to models folder, move tests to test folder
