+++ test/models/transaction.test.ts
@@ -1,7 +1,6 @@
-import TransactionModel, { TransactionStatus } from './transaction';
-import { MongoClient, ObjectId } from 'mongodb';
-import * as mongo from '../../mongo';
-import '../../env-test';
+import TransactionModel, { TransactionStatus } from '../../src/models/transaction';
+import { ObjectId } from 'mongodb';
+import * as mongo from '../../src/mongo';
 
 beforeAll(async () => {
   await mongo.setupConnections();
@@ -13,7 +12,7 @@ describe('Transaction model', () => {
       workspaceId: new ObjectId('5edd36fbb596d4759beb89f6'),
       amount: 100,
       dtCreated: Date.now(),
-      status: TransactionStatus.PENDING,
+      status: TransactionStatus.CONFIRMED,
       userId: new ObjectId('5eb9034a1ccc4421e2623dc2'),
       cardPan: '4455',
     };
@@ -34,6 +33,8 @@ describe('Transaction model', () => {
 });
 
 afterAll(async done => {
-  // Closing the DB connection allows Jest to exit successfully.
+  await mongo.mongoClients.hawk?.close();
+  await mongo.mongoClients.events?.close();
+
   done();
 });
